requirements.txt
# ClinicalDSS Requirements

# LangGraph stack (Path A)
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-chroma>=0.1.2
chromadb>=0.5.5

# Azure OpenAI
openai>=1.35.0

# Azure AI Foundry (Path B)
azure-ai-projects>=1.0.0b1
azure-identity>=1.17.0

# NeMo Guardrails
nemoguardrails>=0.9.0

# Evaluation & Visualization
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0

# API server for UI
fastapi>=0.115.0
uvicorn>=0.30.0
